Added my app project
This is my app 'Quiq Look' - in this assignment I had to develop an app that in some way uses geographical services. This apps function is to let the user report observations (eg. bird watchers) while they are on the go. The submitted observation is then stored with a user added description and geographical position. The user then have the possibility to interact with a map interface to view historic observations. The user can by clicking on icons on a map get details on that observation.

You can access and test the app by using Android Studio 64.